requirements.txt
Flask==2.3.3
Flask-Cors
Flask-Login
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
pymongo==4.6.1
Werkzeug==2.3.7